+++ resources/cluster_role.rb
@@ -2,7 +2,7 @@
 # Cookbook:: sensu-go
 # Resource:: cluster_role
 #
-# Copyright:: 2019 Sensu, Inc.
+# Copyright:: 2020 Sensu, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -27,6 +27,7 @@ include SensuCookbook::SensuMetadataProperties
 include SensuCookbook::SensuCommonProperties
 
 resource_name :sensu_cluster_role
+provides :sensu_cluster_role
 
 # rubocop:disable Style/TrailingCommaInHashLiteral
 property :rules, Array, required: true, callbacks: {