+++ resources/entity.rb
@@ -1,8 +1,8 @@
 #
 # Cookbook:: sensu-go
-# Resource:: agent
+# Resource:: entity
 #
-# Copyright:: 2018 Sensu, Inc.
+# Copyright:: 2020 Sensu, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -27,13 +27,20 @@ include SensuCookbook::SensuMetadataProperties
 include SensuCookbook::SensuCommonProperties
 
 resource_name :sensu_entity
+provides :sensu_entity
 
 action_class do
   include SensuCookbook::Helpers
 end
 
-property :subscriptions, Array
+property :deregister, [true, false]
+property :deregistration, Hash
 property :entity_class, String
+property :redact, Array
+property :sensu_agent_version, String
+property :subscriptions, Array
+property :system, Hash
+property :user, String
 
 action :create do
   directory object_dir do