+++ resources/filter.rb
@@ -2,7 +2,7 @@
 # Cookbook:: sensu-go
 # Resource:: filter
 #
-# Copyright:: 2018 Sensu, Inc.
+# Copyright:: 2020 Sensu, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -27,6 +27,7 @@ include SensuCookbook::SensuMetadataProperties
 include SensuCookbook::SensuCommonProperties
 
 resource_name :sensu_filter
+provides :sensu_filter
 
 property :filter_action, String, equal_to: %w(allow deny), required: true
 property :expressions, Array, required: true