+++ test/cookbooks/sensu_test/recipes/default.rb
@@ -40,6 +40,7 @@ assets.each do |name, property|
   sensu_asset name do
     url property['url']
     sha512 property['checksum']
+    namespace property['namespace']
   end
 end
 
@@ -145,10 +146,29 @@ sensu_mutator 'example-mutator' do
 end
 
 sensu_entity 'example-entity' do
-  subscriptions ['example-entity']
   entity_class 'proxy'
+  subscriptions ['example-entity']
   labels(environment: 'production', region: 'us-west-2')
   annotations(runbook: 'https://www.xkcd.com/378/')
+  redact ['snmp_community_string']
+  system(
+    'hostname': 'example-hypervisor',
+    'platform': 'Citrix Hypervisor',
+    'platform_version': '8.1.0',
+    'network': {
+      'interfaces': [
+        {
+          'name': 'lo',
+          'addresses': ['127.0.0.1/8'],
+        },
+        {
+          'name': 'xapi0',
+          'mac': '52:54:00:20:1b:3c',
+          'addresses': ['172.0.1.72/24'],
+        },
+      ],
+    }
+  )
 end
 
 sensu_hook 'restart_cron_service' do
@@ -181,4 +201,53 @@ sensu_cluster_role_binding 'cluster_admins-all_access' do
   role_name 'all_access'
   role_type 'ClusterRole'
   subjects [ { name: 'cluster-admins', type: 'Group' } ]
+end
+
+sensu_active_directory 'example-active-directory' do
+  ad_servers [{
+    'host': '127.0.0.1',
+    'group_search': {
+      'base_dn': 'dc=acme,dc=org',
+    },
+    'user_search': {
+      'base_dn': 'dc=acme,dc=org',
+    },
+  }]
+end
+
+sensu_active_directory 'example-active-directory-alias' do
+  servers [{
+    'host': '127.0.0.1',
+    'group_search': {
+      'base_dn': 'dc=acme,dc=org',
+    },
+    'user_search': {
+      'base_dn': 'dc=acme,dc=org',
+    },
+  }]
+end
+
+sensu_secrets_provider 'vault' do
+  provider_type 'VaultProvider'
+  address 'https://vaultserver.example.com:8200'
+  max_retries 2
+  rate_limiter(
+    'limit': 10,
+    'burst': 100
+  )
+  timeout '60s'
+  token 'yourVaultToken'
+  version 'v1'
+end
+
+sensu_secret 'env-secret' do
+  namespace 'test-org'
+  id 'CONSUL_TOKEN'
+  secrets_provider 'env'
+end
+
+sensu_secret 'vault-secret' do
+  namespace 'test-org'
+  id 'secret/consul#token'
+  secrets_provider 'vault'
 end